+++ pheasant/lib/router.dart
@@ -1,0 +1,4 @@
+library router;
+
+export 'package:pheasant_temp/pheasant_build.dart'
+  show Route, RouteBase, Router, GlobalRoute;