+++ pheasant_cli/lib/src/constants/buildfile.dart
@@ -46,6 +46,15 @@ String genBuildFile(PheasantCliBaseConfig pheasantConfig,
         },
         'build_to': 'cache',
         'auto_apply': 'root_package'
+      },
+      'pheasantRouterGenerator': {
+        'import': "package:pheasant_build/src/build/builder.dart",
+        'builder_factories': ["pheasantRouterGenerator"],
+        'build_extensions': {
+          ".dart": [".routes.dart"]
+        },
+        'build_to': 'cache',
+        'auto_apply': 'root_package'
       }
     }
   };